Sync daily goal with utility function
